Update windows git detection
